+++ spec/active_item/batch_find_unprocessed_keys_spec.rb
@@ -7,7 +7,7 @@ RSpec.describe 'ActiveItem batch_find unprocessed keys' do
 
   let(:model_class) do
     Class.new(ActiveItem::Base) do
-      self.table_name = 'test-dev-widgets'
+      self.table_name = "#{TABLE_PREFIX}-widgets"
       attr_accessor :name
 
       def self.name
@@ -17,8 +17,8 @@ RSpec.describe 'ActiveItem batch_find unprocessed keys' do
   end
 
   it 'retries unprocessed keys with exponential backoff' do
-    dynamo_client.put_item(table_name: 'test-dev-widgets', item: { 'id' => 'w1', 'name' => 'Alpha' })
-    dynamo_client.put_item(table_name: 'test-dev-widgets', item: { 'id' => 'w2', 'name' => 'Beta' })
+    dynamo_client.put_item(table_name: "#{TABLE_PREFIX}-widgets", item: { 'id' => 'w1', 'name' => 'Alpha' })
+    dynamo_client.put_item(table_name: "#{TABLE_PREFIX}-widgets", item: { 'id' => 'w2', 'name' => 'Beta' })
 
     call_count = 0
     allow(dynamo_client).to receive(:batch_get_item).and_wrap_original do |method, params|
@@ -26,8 +26,8 @@ RSpec.describe 'ActiveItem batch_find unprocessed keys' do
       if call_count == 1
         # Simulate throttling: return w1 but mark w2 as unprocessed
         Aws::DynamoDB::Types::BatchGetItemOutput.new(
-          responses: { 'test-dev-widgets' => [{ 'id' => 'w1', 'name' => 'Alpha' }] },
-          unprocessed_keys: { 'test-dev-widgets' => { keys: [{ 'id' => 'w2' }] } }
+          responses: { "#{TABLE_PREFIX}-widgets" => [{ 'id' => 'w1', 'name' => 'Alpha' }] },
+          unprocessed_keys: { "#{TABLE_PREFIX}-widgets" => { keys: [{ 'id' => 'w2' }] } }
         )
       else
         method.call(params)
@@ -44,8 +44,8 @@ RSpec.describe 'ActiveItem batch_find unprocessed keys' do
   it 'gives up after max retries' do
     allow(dynamo_client).to receive(:batch_get_item).and_return(
       Aws::DynamoDB::Types::BatchGetItemOutput.new(
-        responses: { 'test-dev-widgets' => [] },
-        unprocessed_keys: { 'test-dev-widgets' => { keys: [{ 'id' => 'w1' }] } }
+        responses: { "#{TABLE_PREFIX}-widgets" => [] },
+        unprocessed_keys: { "#{TABLE_PREFIX}-widgets" => { keys: [{ 'id' => 'w1' }] } }
       )
     )
     allow_any_instance_of(Object).to receive(:sleep)